% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.year.R
\name{as.year}
\alias{-.year}
\alias{as.year}
\alias{as.year.Date}
\alias{as.year.default}
\alias{is.year}
\alias{width.year}
\title{Year vectors}
\usage{
as.year(x)

\method{as.year}{Date}(x)

\method{as.year}{default}(x)

is.year(x)

\method{-}{year}(a, b)

width.year(x)
}
\arguments{
\item{x}{object to be coerced or tested}

\item{a, b}{years to be added or subtracted}
}
\value{
Vector of class \code{year} and "AsIs" (see function \code{\link{I}}).
}
\description{
Creates or coerces objects of type \code{year}.
}
\details{
The year class also inherits methods from the character class but not from
numeric. (It does not make any sence to, for example, multiply two years with
each other). There are however methods for subtraction and it is also
possible to add a integer to a year etcetera.
}
\examples{
as.year("2012")
as.year(Sys.Date())
rccmisc::width(c(2012, 2014))
}

