% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asDate.R
\name{asDate}
\alias{asDate}
\title{as.Date with format = "\%Y-\%m-\%d"}
\usage{
asDate(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{as.Date}}}
}
\description{
This is just a wrapper for the standard INCA date format
}
\examples{
asDate(c("", "", "2014-10-10"))
}

